Move createPncDisposalsFromResult into root lib folder
As it's used by Phase 2 and 3.
