+++ packages/core/lib/createPncDisposalsFromResult/createPncDisposal.test.ts
