+++ packages/core/lib/createPncDisposalsFromResult/formatDateSpecifiedInResult.ts
