+++ packages/core/lib/createPncDisposalsFromResult/getFirstDateSpecifiedInResult.test.ts
@@ -1,6 +1,6 @@
-import type { Result } from "../../../types/AnnotatedHearingOutcome"
+import type { Result } from "../../types/AnnotatedHearingOutcome"
 
-import ResultClass from "../../../types/ResultClass"
+import ResultClass from "../../types/ResultClass"
 import getFirstDateSpecifiedInResult from "./getFirstDateSpecifiedInResult"
 
 const acceptedResultClasses = [
