+++ packages/core/lib/createPncDisposalsFromResult/index.ts
