+++ packages/core/lib/createPncDisposalsFromResult/isAmountSpecifiedInResultValid.test.ts
