+++ packages/core/lib/createPncDisposalsFromResult/isAmountSpecifiedInResultValid.ts
