+++ packages/core/lib/triggers/TRPS0003.ts
@@ -3,9 +3,9 @@ import TriggerCode from "@moj-bichard7-developers/bichard7-next-data/dist/types/
 import type { Trigger } from "../../types/Trigger"
 import type { TriggerGenerator } from "../../types/TriggerGenerator"
 
-import { maxDisposalTextLength } from "../../phase2/lib/createPncDisposalsFromResult/createPncDisposalByFirstAndSecondDurations"
 import { getDisposalTextFromResult } from "../../phase2/lib/getDisposalTextFromResult"
 import Phase from "../../types/Phase"
+import { maxDisposalTextLength } from "../createPncDisposalsFromResult/createPncDisposalByFirstAndSecondDurations"
 import forEachRecordableResult from "../forEachRecordableResult"
 
 const triggerCode = TriggerCode.TRPS0003
