+++ packages/core/phase2/exceptions/HO200205.ts
@@ -4,9 +4,9 @@ import type { AnnotatedHearingOutcome, Result } from "../../types/AnnotatedHeari
 import type Exception from "../../types/Exception"
 import type { ExceptionGenerator } from "../../types/ExceptionGenerator"
 
+import isAmountSpecifiedInResultValid from "../../lib/createPncDisposalsFromResult/isAmountSpecifiedInResultValid"
 import errorPaths from "../../lib/exceptions/errorPaths"
 import forEachRecordableResult from "../../lib/forEachRecordableResult"
-import isAmountSpecifiedInResultValid from "../lib/createPncDisposalsFromResult/isAmountSpecifiedInResultValid"
 
 const firstAmountIndex = 0
 const thirdAmountIndex = 2
