+++ packages/core/phase3/lib/hearingDetails/createAdjudicationFromOffence.ts
@@ -1,8 +1,8 @@
 import type { Offence } from "../../../types/AnnotatedHearingOutcome"
 import type { Adjudication } from "../../types/HearingDetails"
 
+import formatDateSpecifiedInResult from "../../../lib/createPncDisposalsFromResult/formatDateSpecifiedInResult"
 import createPncAdjudicationFromAho from "../../../phase2/lib/createPncAdjudicationFromAho"
-import formatDateSpecifiedInResult from "../../../phase2/lib/createPncDisposalsFromResult/formatDateSpecifiedInResult"
 import { HearingDetailsType } from "../../types/HearingDetails"
 
 export const createAdjudicationFromOffence = (offence: Offence, dateOfHearing: Date): Adjudication | undefined => {
