+++ packages/core/phase3/lib/pncUpdateRequestGenerators/remandGenerator/remandGenerator.ts
@@ -2,8 +2,8 @@ import { isError } from "@moj-bichard7/common/types/Result"
 
 import type PncUpdateRequestGenerator from "../../../types/PncUpdateRequestGenerator"
 
+import formatDateSpecifiedInResult from "../../../../lib/createPncDisposalsFromResult/formatDateSpecifiedInResult"
 import { lookupRemandStatusByCjsCode } from "../../../../lib/dataLookup"
-import formatDateSpecifiedInResult from "../../../../phase2/lib/createPncDisposalsFromResult/formatDateSpecifiedInResult"
 import { PncOperation } from "../../../../types/PncOperation"
 import addPaddingToBailCondition from "../../addPaddingToBailCondition"
 import generateBasePncUpdateRequest from "../../generateBasePncUpdateRequest"
